requirements-dev.txt
# Development requirements for Python 3.X
# For package requirements, please see setup.py

pytest==6.2.4
pytest-cov==2.12.0
pytest-django==4.5.2
pytest-mock~=3.6.1
tox==3.23.1
wheel==0.36.2
factory-boy~=3.2.0
freezegun~=1.2.1
smartevents  --index-url=${PIP_EXTRA_INDEX_URL}
ipdb
